Dev 3.6 1. Выполнено задание на 13 спринт
С уважением Жуков Игорь

+++ app/src/main/java/rs/example/playlistmaker/utils/StaffFunctions.kt
@@ -5,7 +5,7 @@ import java.util.Locale
 
 object StaffFunctions {
 
-    fun getSimpleDateFormat(i: Int): String {
+    fun getSimpleDateFormat(i: Long): String {
         return SimpleDateFormat(
             "mm:ss",
             Locale.getDefault()
